fix: address PR #4056 review comments in xmlParser.cpp
- Add <stddef.h> and <limits.h> to fix SIZE_MAX/INT_MAX undeclared on Clang
- Roll back staged memory in addChild/addAttribute/addText/addClear when
  addToOrder() fails, preventing memory leaks on OOM paths
- Convert CreateXMLStringR accumulators (nResult, cb, cbElement) and return
  type from int to size_t to prevent signed integer overflow on large documents
- Guard toXMLStringFast size cast with INT_MAX check before narrowing size_t to int

+++ buildroot-external/package/libxmlparser/xmlParser.h
@@ -245,7 +245,7 @@ private:
     int ParseClearTag(void *pXML, void *pClear);
     int ParseXMLElement(void *pXML);
     bool addToOrder(int index, int type);
-    static int CreateXMLStringR(XMLNodeData *pEntry, LPTSTR lpszMarker, int nFormat);
+    static size_t CreateXMLStringR(XMLNodeData *pEntry, LPTSTR lpszMarker, int nFormat);
     static void *enumContent(XMLNodeData *pEntry,int i, XMLElementType *nodeType);
     static int nElement(XMLNodeData *pEntry);
     static void removeOrderElement(XMLNodeData *d, XMLElementType t, int index);
